Resolve conflict in src/png/filter.rs by keeping AdaptiveFast in
parallel path condition for proper parallelization of adaptive filters.

+++ examples/png_timing.rs
@@ -15,14 +15,8 @@ fn main() {
 
     let mut args = env::args().skip(1);
     let pattern = args.next().unwrap_or_else(|| "gradient".to_string());
-    let width: u32 = args
-        .next()
-        .and_then(|s| s.parse().ok())
-        .unwrap_or(512);
-    let height: u32 = args
-        .next()
-        .and_then(|s| s.parse().ok())
-        .unwrap_or(512);
+    let width: u32 = args.next().and_then(|s| s.parse().ok()).unwrap_or(512);
+    let height: u32 = args.next().and_then(|s| s.parse().ok()).unwrap_or(512);
 
     let bytes_per_pixel = 4; // RGBA
     let mut pixels = vec![0u8; width as usize * height as usize * bytes_per_pixel];
@@ -34,26 +28,16 @@ fn main() {
     let options = PngOptions::fast();
 
     let t_filter = Instant::now();
-    let filtered = filter::apply_filters(
-        &pixels,
-        width,
-        height,
-        bytes_per_pixel as usize,
-        &options,
-    );
+    let filtered = filter::apply_filters(&pixels, width, height, bytes_per_pixel, &options);
     let filter_time = t_filter.elapsed();
 
     let t_deflate = Instant::now();
-    let (compressed, stats) =
-        deflate_zlib_packed_with_stats(&filtered, options.compression_level);
+    let (compressed, stats) = deflate_zlib_packed_with_stats(&filtered, options.compression_level);
     let deflate_time = t_deflate.elapsed();
 
     println!("Pattern: {pattern}, size: {width}x{height}, bpp: {bytes_per_pixel}");
     println!("Filtered bytes: {}", filtered.len());
-    println!(
-        "Filter time: {:.3} ms",
-        filter_time.as_secs_f64() * 1000.0
-    );
+    println!("Filter time: {:.3} ms", filter_time.as_secs_f64() * 1000.0);
     println!(
         "Deflate total: {:.3} ms (lz77 {:.3} ms, fixed {:.3} ms, dynamic {:.3} ms, choose {:.3} ms)",
         deflate_time.as_secs_f64() * 1000.0,
